Fixed database schema for users

+++ Backend/user-service/src/main/resources/db/migration/V1__create_schema_.sql
@@ -20,7 +20,7 @@ CREATE TABLE IF NOT EXISTS Users (
   Address VARCHAR(255) NOT NULL,
   Notifications BOOLEAN NOT NULL,
   CardName VARCHAR(255),
-  CardNumber INT,
+  CardNumber BIGINT,
   CardExpiration VARCHAR(255),
   CardCVV INT,
   CONSTRAINT UQ_Username UNIQUE (Username),
